+++ okitweb/static/okit/properties/js/okit_properties.js
@@ -563,7 +563,6 @@ class OkitResourceProperties {
         if (empty_option) select.append('option').attr('value', '').attr('selected', 'selected').text(empty_value ? empty_value : '')
         let id = ''
         const resources = okitOciData[resource_type](filter)
-        console.info('Resources', resource_type, resources, okitOciData[resource_type])
         if (groups) {
             Object.entries(groups).forEach(([k, v]) => {
                 console.info('Group', k, 'Filter', v)
@@ -578,7 +577,7 @@ class OkitResourceProperties {
                 })
             })
         } else {
-            resources.forEach((r, i) => {
+            resources.sort((a, b) => a[display_element].localeCompare(b[display_element])).forEach((r, i) => {
                 r = r instanceof Object ? r : {id: r, display_name: r}
                 const option = select.append('option').attr('value', r[id_element]).text(r[display_element])
                 if (!empty_option && i === 0) {